1、三分钟没有定位记录结束线路 2、速度小于7km/h（可配置）时持续120s（可配置）的线路调用结束算分

+++ libWildebeest/src/main/java/com/jxtii/wildebeest/core/AMAPLocalizer.java
@@ -213,11 +213,13 @@ public class AMAPLocalizer implements AMapLocationListener {
                 logAndWrite("isStart=" + isStart + ";curSpeed=" + curSpeed + ";beginSpeed=" + CommUtil.BEGIN_SPEED + ";" + (curSpeed > CommUtil.BEGIN_SPEED), LogEnum.INFO, true);
                 if (curSpeed > CommUtil.BEGIN_SPEED && !isStart) {
                     uploadInitInfo();
-                } else if (curSpeed != 0.0) {
+                } else if (curSpeed != 0.0 && CommUtil.END_SPEED <= curSpeed) {
                     DataSupport.deleteAll(NoGpsInfo.class);
+                } else if (curSpeed < CommUtil.END_SPEED) {
+                    validNoGpsInfo();
                 } else if (curSpeed == 0.0) {
                     validNoGpsInfo();
-                }else {
+                } else {
                     logAndWrite("other situation", LogEnum.INFO, true);
                 }
             } else {
